Set up Comment Controller with nested BP
