+++ controllers/product_controller.py
@@ -4,12 +4,11 @@ from models.product import Product, product_schema, products_schema
 from sqlalchemy.exc import IntegrityError
 from psycopg2 import errorcodes
 from flask_jwt_extended import get_jwt_identity, jwt_required
-from models.comment import Comment, comment_schema, comments_schema
-from models.order import Order, order_schema, orders_schema
-from datetime import date
+from controllers.comment_controller import comments_bp
+
 
 products_bp = Blueprint('products', __name__, url_prefix = '/products')
-
+products_bp.register_blueprint(comments_bp)
 
 @products_bp.route('/')
 def get_products():
@@ -77,59 +76,3 @@ def update_product(id):
         return {'error': f'Product with id {id} not found.'}, 404
     
 
-
-
-
-
-
-
-
-# Adding Comment controllers to Products controller as comments only exist on a Product entity
-
-@products_bp.route('/<int:product_id>/comments', methods = ['POST'])
-@jwt_required()
-def create_comment(product_id):
-    body_data = request.get_json()
-    qry = db.select(Product).filter_by(id = product_id) 
-    product = db.session.scalar(qry)
-    if product:
-        comment = Comment(
-            message = body_data.get('message'),
-            user_id = get_jwt_identity(),
-            product = product
-        )
-        db.session.add(comment)
-        db.session.commit()
-        return comment_schema.dump(comment), 201
-    else:
-        return {'error': f'Product not found with id {product_id}.'}, 404
-        
-
-# No GET method for comments, all comments will exist and be related to products, so any GET product will display comments 
-
-
-@products_bp.route('/<int:product_id>/comments/<int:comment_id>', methods = ['DELETE'])
-@jwt_required()
-def delete_comment(product_id, comment_id):
-    qry = db.select(Comment).filter_by(id = comment_id)
-    comment = db.session.scalar(qry)
-    if comment:
-        db.session.delete(comment)
-        db.session.commit()
-        return {'message': f'Comment "{comment.message}" has been deleted successfully.'}
-    else: 
-        return {'error': f'Comment with id {comment_id} not found.'}, 404
-    
-
-@products_bp.route('/<int:product_id>/comments/<int:comment_id>', methods = ['PUT', 'PATCH'])
-@jwt_required()
-def edit_comment(product_id, comment_id):
-    body_data = request.get_json()
-    qry = db.select(Comment).filter_by(id = comment_id)
-    comment = db.session.scalar(qry)
-    if comment:
-        comment.message = body_data.get('message') or comment.message
-        db.session.commit()
-        return comment_schema.dump(comment)
-    else:
-        return {'error': f'Comment with id {comment_id} not found.'}
